move initialization test base functinality tests to base controller unit test
