+++ tests/unit/test_init.py
@@ -7,7 +7,6 @@ from xrdsst.models.initialization_status import InitializationStatus
 from xrdsst.rest.rest import ApiException
 import yaml
 
-
 class TestInit(unittest.TestCase):
 
     def __init__(self, *args, **kwargs):
@@ -77,51 +76,6 @@ class TestInit(unittest.TestCase):
             init = InitServerController()
             init.init_security_server(self._config, self._ss_config["security-server"][0])
             self.assertRaises(ApiException)
-
-    def test_load_config(self):
-        init = InitServerController()
-        temp_file_name = "base.yaml"
-        config_file = open(temp_file_name, "w")
-        config_file.close()
-        with open(temp_file_name, "r") as yml_file:
-            cfg = yaml.load(yml_file, Loader=yaml.FullLoader)
-        response = init.load_config(temp_file_name)
-        os.remove(temp_file_name)
-        assert response == cfg
-
-    def test_load_config_exception(self):
-        init = InitServerController()
-        config_file = "conf.yaml"
-        self.assertRaises(FileNotFoundError, lambda: init.load_config(config_file))
-
-    def test_init_logging(self):
-        temp_file_name = "temp.log"
-        log_file = open(temp_file_name, "w")
-        self._ss_config["logging"][0]["file"] = temp_file_name
-        init = InitServerController()
-        response = init.init_logging(self._ss_config)
-        log_file.close()
-        os.remove(temp_file_name)
-        self.assertEqual(response, None)
-
-    def test_init_logging_exception(self):
-        temp_file_name = "temp.log"
-        self._ss_config["logging"][0]["file"] = temp_file_name
-        init = InitServerController()
-        init.init_logging(self._ss_config)
-        self.assertRaises(FileNotFoundError)
-
-    def test_initialize_basic_conf_values(self):
-        init = InitServerController()
-        security_server = self._ss_config["security-server"][0]
-        configuration = Configuration()
-        configuration.api_key['Authorization'] = security_server["api_key"]
-        configuration.host = security_server["url"]
-        configuration.verify_ssl = False
-        response = init.initialize_basic_config_values(security_server)
-        assert response.api_key == configuration.api_key
-        assert response.host == configuration.host
-        assert response.verify_ssl == configuration.verify_ssl
 
     def test_initialize_server_when_already_initialized(self):
         initialization_status = InitializationStatus(is_anchor_imported=True,
